requirements.txt
langchain>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.18
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
